[IMP] l10n_hu_edi: merge l10n_hu_edi_receive into l10n_hu_edi
Merge `l10n_hu_edi_receive` into `l10n_hu_edi`.

Upgrade PR: https://github.com/odoo/upgrade/pull/10626

task-6332269
