+++ addons/l10n_hu_edi/data/template_requests.xml
@@ -94,4 +94,32 @@
         </InvoiceAnnulment>
     </template>
 
+    <template id="query_invoice_digest_request">
+        <QueryInvoiceDigestRequest xmlns:common="http://schemas.nav.gov.hu/NTCA/1.0/common" xmlns="http://schemas.nav.gov.hu/OSA/3.0/api">
+            <t t-call="l10n_hu_edi.request_header"/>
+            <page t-out="page"/>
+            <invoiceDirection t-out="invoiceDirection"/>
+            <invoiceQueryParams>
+                <mandatoryQueryParams>
+                    <insDate>
+                        <dateTimeFrom t-out="dateTimeFrom"/>
+                        <dateTimeTo t-out="dateTimeTo"/>
+                    </insDate>
+                </mandatoryQueryParams>
+            </invoiceQueryParams>
+        </QueryInvoiceDigestRequest>
+    </template>
+
+    <template id="query_invoice_data_request">
+        <QueryInvoiceDataRequest xmlns:common="http://schemas.nav.gov.hu/NTCA/1.0/common" xmlns="http://schemas.nav.gov.hu/OSA/3.0/api">
+            <t t-call="l10n_hu_edi.request_header"/>
+            <invoiceNumberQuery>
+                <invoiceNumber t-out="invoiceNumber"/>
+                <invoiceDirection t-out="invoiceDirection"/>
+                <batchIndex t-if="batchIndex" t-out="batchIndex"/>
+                <supplierTaxNumber t-out="supplierTaxNumber"/>
+            </invoiceNumberQuery>
+        </QueryInvoiceDataRequest>
+    </template>
+
 </odoo>
