+++ addons/l10n_hu_edi/security/ir.access.csv
@@ -1,3 +1,4 @@
 id,name,model_id,group_id/id,operation,domain
 l10n_hu_edi_cancellation_user,l10n_hu_edi.cancellation group_account_invoice,l10n_hu_edi.cancellation,account.group_account_invoice,crud,
 l10n_hu_edi_tax_audit_export_user,l10n_hu_edi.tax_audit_export group_account_manager,l10n_hu_edi.tax_audit_export,account.group_account_manager,crud,
+access_l10n_hu_edi_receive_bills_wizard,l10n_hu_edi.receive.bills.wizard,l10n_hu_edi.receive.bills.wizard,account.group_account_invoice,crud,
