+++ addons/l10n_hu_edi/static/src/views/sync_with_nav_btn/sync_with_nav_btn.xml
@@ -1,12 +1,12 @@
 <templates>
 
-    <t t-name="l10n_hu_edi_receive.ListView.Buttons" t-inherit="account.AccountMoveListView.Buttons" t-inherit-mode="primary">
+    <t t-name="l10n_hu_edi.ListView.Buttons" t-inherit="account.AccountMoveListView.Buttons" t-inherit-mode="primary">
         <xpath expr="." position="inside">
             <SyncWithNavBtn/>
         </xpath>
     </t>
 
-    <t t-name="l10n_hu_edi_receive.SyncWithNavBtn">
+    <t t-name="l10n_hu_edi.SyncWithNavBtn">
         <t t-if="this.showButton">
             <button class="btn btn-secondary" t-on-click="this.openWizard">Sync with NAV</button>
         </t>
