+++ addons/l10n_hu_edi/tests/__init__.py
@@ -1,8 +1,9 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import common
+from . import test_account_move_sync_tax_lines
+from . import test_bill_xml_upload
 from . import test_credit_debit_notes
 from . import test_flows_live
 from . import test_flows_mocked
 from . import test_invoice_xml
-from . import test_account_move_sync_tax_lines
