+++ addons/l10n_hu_edi/tests/invoice_xmls/credit_note_original_bill.xml
