+++ addons/l10n_hu_edi/tests/invoice_xmls/debit_note_original_bill.xml
