+++ addons/l10n_hu_edi/tests/invoice_xmls/normal_invoice.xml
