+++ addons/l10n_hu_edi/tests/invoice_xmls/query_invoice_data_response_gzip.xml
