+++ addons/l10n_hu_edi/tests/invoice_xmls/simplified_invoice.xml
