+++ addons/l10n_hu_edi/tests/mocked_requests/BATCH_MOD_2026_0002_1.xml
