+++ addons/l10n_hu_edi/tests/mocked_requests/queryInvoiceData_request.xml
