+++ addons/l10n_hu_edi/tests/mocked_requests/queryInvoiceDigest_request.xml
