+++ addons/l10n_hu_edi/views/account_move_views.xml
@@ -11,6 +11,17 @@
         </field>
     </record>
 
+    <record id="l10n_hu_edi_view_in_invoice_bill_tree" model="ir.ui.view">
+        <field name="name">l10n_hu_edi.view_in_invoice_bill_tree</field>
+        <field name="model">account.move</field>
+        <field name="inherit_id" ref="account.view_in_invoice_bill_tree"/>
+        <field name="arch" type="xml">
+            <list position="attributes">
+                <attribute name="js_class">l10n_hu_edi_account_move_tree</attribute>
+            </list>
+        </field>
+    </record>
+
     <record id="view_move_form_inherit_l10n_hu_edi" model="ir.ui.view">
         <field name="name">account.move.form.inherit.l10n_hu_edi</field>
         <field name="model">account.move</field>
