+++ addons/l10n_hu_edi/wizard/__init__.py
@@ -1,4 +1,5 @@
 from . import account_move_debit
 from . import account_move_reversal
 from . import l10n_hu_edi_cancellation
+from . import l10n_hu_edi_receive_bills_wizard
 from . import l10n_hu_edi_tax_audit_export
