+++ addons/l10n_hu_edi/wizard/l10n_hu_edi_receive_bills_wizard_views.xml
@@ -2,13 +2,13 @@
 <odoo>
 
     <record id="l10n_hu_edi_receive_bills_wizard_form" model="ir.ui.view">
-        <field name="name">l10n_hu_edi_receive.bills.wizard.form</field>
-        <field name="model">l10n_hu_edi_receive.bills.wizard</field>
+        <field name="name">l10n_hu_edi.receive.bills.wizard.form</field>
+        <field name="model">l10n_hu_edi.receive.bills.wizard</field>
         <field name="arch" type="xml">
             <form>
                 <group>
-                    <field name="l10n_hu_edi_receive_from" string="Datetime range" widget="daterange" options="{'end_date_field': 'l10n_hu_edi_receive_to'}"/>
-                    <field name="l10n_hu_edi_receive_to" invisible="1"/>  <!-- needed for daterange widget -->
+                    <field name="datetime_from" string="Datetime range" widget="daterange" options="{'end_date_field': 'datetime_to'}"/>
+                    <field name="datetime_to" invisible="1"/>  <!-- needed for daterange widget -->
                 </group>
                 <footer>
                     <button name="action_receive_bills" string="Fetch Bills"
