--- addons/l10n_hu_edi_receive/__init__.py
@@ -1,4 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import models
-from . import wizard
