--- addons/l10n_hu_edi_receive/models/__init__.py
@@ -1,5 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import account_move
-from . import l10n_hu_edi_connection
-from . import res_company
