--- addons/l10n_hu_edi_receive/security/ir.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id,group_id/id,operation,domain
-access_l10n_hu_edi_receive_bills_wizard,l10n_hu_edi_receive.bills.wizard,l10n_hu_edi_receive.bills.wizard,account.group_account_invoice,crud,
