--- addons/l10n_hu_edi_receive/tests/__init__.py
@@ -1,5 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import common
-from . import test_api_mocked
-from . import test_bill_xml_upload
