--- addons/l10n_hu_edi_receive/wizard/__init__.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import l10n_hu_edi_receive_bills_wizard
